+++ EliteAPI/Models/DTOs/Outgoing/Player.cs
@@ -14,7 +14,7 @@ public class PlayerDataDto
     public long LastLogout { get; set; }
 
     public int Karma { get; set; }
-    public int NetworkExp { get; set; }
+    public double NetworkExp { get; set; }
    
     public int RewardHighScore { get; set; }
     public int RewardScore { get; set; }
